+++ lib/zarby.rb
@@ -4,10 +4,11 @@ require_relative 'zarby/version'
 require_relative 'zarby/csv'
 require_relative 'zarby/normalize'
 require_relative 'zarby/anonymize'
+require_relative 'zarby/errors'
 
 module Zarby
-  def anonymise(input, nb_chars: 2, mask_with: '*')
-    Anonymize.new(input:, nb_chars: nb_chars, mask_with: mask_with).call
+  def anonymize(input, nb_chars: 2, mask_with: '*')
+    Anonymize.new(input:, nb_chars:, mask_with:).call
   end
 
   def detect_separator(content)
@@ -18,6 +19,6 @@ module Zarby
     Normalize.new(input:).call
   end
 
-  module_function :anonymise, :detect_separator, :utf8 
+  module_function :anonymize, :detect_separator, :utf8 
   private_constant :Anonymize
 end