+++ lib/zarby/errors.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module Zarby
+  class ColSeparatorZarby < StandardError; end
+  class ArgumentZarby < StandardError; end
+end