+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 * Poro class Csv and Normalize to symplify syntax
 * Csv and Normalize class will be private and their method depreciated
 * Use only instance of Zarby
-* Add module to anonymise string value.
+* Add module to anonymize string value.
 
 # 0.1.5 / 2023-10-24